BarrierLocationsNov2
In this geodatabase, there is a shapefile containing all the barrier
locations I have identified as of Nov 1. This includes oberservation
numbers 1-30. There are some other shapefiles, like streets, wards, and
parks. There is also a word document that details the conflicts between
the Waldron Data and what I found.