Updated Barriers Shapefile
Here is the most updated barriers shapefile. I have removed the
shapefile of the barriers that I geocoded. The newest shapefile that I
have uploaded is called STL_STRUCTURE_Barriers. This is the shapefile
that should be used from here on. There are three observations that did
not join to a proper ward. Their barrier id numbers are 143, 216, and
229.